requirements.txt
Flask
flask-cors
gunicorn
SQLAlchemy
waitress
psycopg2-binary
flask-sqlalchemy
python-dotenv
PyJWT
bcrypt



